Updated Sonicwall filters and Dashboards (#18)
Updated sonicwall filter to support the timezone in the timestamp, which is required when using a GMS to manage the sonicwall cluster. Also updated the syslog sonicwall test to reflect the additional log variation.

Updated sonicwall syslog filter to only apply additional filters to data tagged with "sonicwall_syslog" and also enabled geo tagging

Updated sonicwall dashboards with graphs for:
 * Blocked websites bar chart 
 * Blocked website tables
 * Port scan map
 * Port scan chart
 * IP spoof chart
 * Sonicwall IPS attack table.

Added sonicwall elastalerts on IPS Medium and high priority attacks.